+++ crates/teloxide-core/src/adaptors/auto_send.rs
@@ -177,7 +177,6 @@ where
 }
 
 download_forward! {
-    'w
     B
     AutoSend<B>
     { this => this.inner() }