+++ crates/teloxide-core/src/adaptors/throttle/requester_impl.rs
@@ -182,7 +182,6 @@ where
 }
 
 download_forward! {
-    'w
     B
     Throttle<B>
     { this => this.inner() }