+++ crates/teloxide-core/src/lib.rs
@@ -7,7 +7,7 @@
 //!```toml
 //! teloxide_core = "0.9"
 //! ```
-//! _Compiler support: requires rustc 1.64+_.
+//! _Compiler support: requires rustc 1.65+_.
 //!
 //! ```
 //! # async {