+++ crates/teloxide-core/src/net/download.rs
@@ -12,14 +12,17 @@ use tokio::io::{AsyncWrite, AsyncWriteExt};
 use crate::{errors::DownloadError, net::file_url};
 
 /// A trait for downloading files from Telegram.
-// FIXME(waffle): ideally, this lifetime ('w) shouldn't be here, but we can't
-//                help it without GATs
-pub trait Download<'w> {
+pub trait Download {
     /// An error returned from [`download_file`](Self::download_file).
-    type Err;
+    type Err<'dst>;
 
     /// A future returned from [`download_file`](Self::download_file).
-    type Fut: Future<Output = Result<(), Self::Err>> + Send;
+    type Fut<'dst>: Future<Output = Result<(), Self::Err<'dst>>> + Send;
+
+    // NOTE: We currently only allow borrowing `dst` in the future,
+    //       however we could also allow borrowing `self` or `path`.
+    //       This doesn't seem useful for our current implementers of
+    //       `Download`, but we could.
 
     /// Download a file from Telegram into `destination`.
     ///
@@ -49,11 +52,11 @@ pub trait Download<'w> {
     ///
     /// [`GetFile`]: crate::payloads::GetFile
     /// [`download_file_stream`]: Self::download_file_stream
-    fn download_file(
+    fn download_file<'dst>(
         &self,
         path: &str,
-        destination: &'w mut (dyn AsyncWrite + Unpin + Send),
-    ) -> Self::Fut;
+        destination: &'dst mut (dyn AsyncWrite + Unpin + Send),
+    ) -> Self::Fut<'dst>;
 
     /// An error returned from
     /// [`download_file_stream`](Self::download_file_stream).