+++ crates/teloxide-core/CHANGELOG.md
@@ -45,10 +45,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RequestError::RetryAfter` single fields type to `Seconds` ([#859][pr859])
 - `CallbackGame`, `ForumTopicClosed`, `ForumTopicReopened`, `GeneralForumTopicHidden`, `GeneralForumTopicUnhidden` and `WriteAccessAllowed` structures
   are now defined as named (`struct S {}`) instead of unit (`struct S;`) in order to fix their deserialization ([#876][pr876])
+- `Download` now uses GAT feature on the `Fut` and `Err` associated types, instead of a lifetime on the whole trait ([#885][pr885])
+- MSRV (Minimal Supported Rust Version) was bumped from `1.64.0` to `1.65.0`
 
 [pr852]: https://github.com/teloxide/teloxide/pull/853
 [pr859]: https://github.com/teloxide/teloxide/pull/859
 [pr876]: https://github.com/teloxide/teloxide/pull/876
+[pr885]: https://github.com/teloxide/teloxide/pull/885
 
 ### Deprecated